refactor(result): namespace changed
